[B] ADD_ProfileImg select, insert API
#S07P12C202-507
사용자 정보 수정(프로필 사진, 관련정보)하는 api를 설계하였습니다.

+++ backend/src/main/java/com/sos/owo/dto/MemberResponseDto.java
@@ -15,7 +15,7 @@ public class MemberResponseDto {
     int age;
     double height;
     double weight;
-    double activityNum;
-    double activityHour;
+    int activityNum;
+    int activityHour;
 
 }
